fix migration

+++ database/migrations/2023_04_03_145259_create_products_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->id();
             $table->string('title');
             $table->integer('quantity');
-            $table->float('price');
-            $table->float('total_price');
+            $table->unsignedBigInteger('price');
+            $table->unsignedBigInteger('total_price');
             $table->timestamps();
         });
     }
